Extract DomContentSearchClusterBuilder
